Port drawing code to main oled driver
